add cuda support in partition examples

+++ examples/filters/fvector_cpu.c
@@ -20,11 +20,11 @@
 
 void vector_cpu_func(void *buffers[], void *cl_arg)
 {
-        unsigned i;
+        int i;
         int *factor = (int *) cl_arg;
 
         /* length of the vector */
-        unsigned n = STARPU_VECTOR_GET_NX(buffers[0]);
+        int n = (int)STARPU_VECTOR_GET_NX(buffers[0]);
         /* local copy of the vector pointer */
         int *vector = (int *)STARPU_VECTOR_GET_PTR(buffers[0]);
